修复AsyncSerialStream.Read的Number可能超出范围的问题

+++ +Async_stream_IO/AsyncSerialStream.m
@@ -473,7 +473,7 @@ classdef AsyncSerialStream<Async_stream_IO.IAsyncStream
 			for V=1:numel(varargin)
 				Arg=varargin{V};
 				if isnumeric(Arg)
-					Number=Arg;
+					Number=uint16(Arg);
 				else
 					Type=Arg;
 				end
